Extract the spline code into a standalone module
